ffi cleanup: object: initial changes
